requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==2.2.0
djangorestframework==3.8.2
idna==2.7
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
requests==2.19.1
sqlparse==0.3.0
urllib3==1.23
virtualenv==16.6.1
